+++ hash.py
@@ -1,2 +1,0 @@
-from werkzeug.security import generate_password_hash
-print(generate_password_hash("1234"))